Move components to containers & update category filter
