--- src/App.js
@@ -1,7 +1,0 @@
-const App = () => (
-  <div className="App">
-    <h1>Hello from Jurgen and Yahir</h1>
-  </div>
-);
-
-export default App;
